Use uintptr_t for texture_id.
Not all platforms use an uint32 to identify a texture. Use an
uintptr_t so we can safely cast a pointer to a texture_id.

This is for b/20429730.

+++ media/base/video_frame.cc
@@ -313,7 +313,7 @@ uint8* VideoFrame::data(size_t plane) const {
   return data_[plane];
 }
 
-uint32 VideoFrame::texture_id() const {
+uintptr_t VideoFrame::texture_id() const {
   DCHECK_EQ(format_, NATIVE_TEXTURE);
   return texture_id_;
 }
